all in one

+++ build.sh
@@ -2,7 +2,7 @@
 #set -u
 
 display_usage(){
-	echo -e "Usage: build.sh [OPTIONS] [VALUE] \n Options: \n -m : model filename, format onnx. \n -t : build target, llvm or cuda, default llvm."	
+	echo -e "Usage: build.sh [OPTIONS] [VALUE] \n Options: \n -m : model filename, format onnx. \n -t : build target, llvm or cuda, default llvm.  \n -c : camera address. \n -o : output path."	
 } 
 
 if [ -z $1 ]; then 
@@ -10,7 +10,9 @@ if [ -z $1 ]; then
 	exit 0
 fi
 target=llvm
-while getopts "m:t:f:h?" opt
+camera_address=rtsp://192.168.1.59/video
+output_path=./output
+while getopts "m:t:c:o:h?" opt
 do
 	case ${opt} in
 		m)
@@ -19,8 +21,11 @@ do
 		t)
 		target=$OPTARG
 		;;
-		f)
-		filename=$OPTARG
+		c)
+		camera_address=$OPTARG
+		;;
+		o)
+		output_path=$OPTARG
 		;;
 		*)
 		display_usage
@@ -35,13 +40,38 @@ if [ -z $model ]; then
 fi
 echo "use model ${model}"
 echo "use target ${target}"
-echo "inference file is ${filename}"
 
-docker run --rm -v $(pwd):/workspace phlens/tvm_py_build:arm64 --model ${model} --target ${target}
+model_prefix=${model%.*}
+model_run=$model_prefix".so"
+
+if [ ! -f ${model_run} ]; then #model is not built
+	echo "model lib is not built, start building------!"
+	docker run --rm -v $(pwd):/workspace phlens/tvm_py_build:arm64 --model ${model} --target ${target}
+	if [ $? -ne 0 ]; then
+		echo "build failed : "$?
+		exit 1
+	else
+		echo "finish build model!"
+	fi
+fi
+
+if [ ${target} = llvm ]; then
+	device=True
+	echo "use cpu inference!"
+else
+	device=False
+	echo "use gpu inference!"
+fi
+
+echo "start service----------"
+echo "use model lib ${model_run}"
+
+docker run -d -w /workspace -v $(pwd):/workspace phlens/tvm_opencv_runtime:arm64 python3 little_model.py --model ${model_run} --cpu ${device} --camera_address ${camera_address} --output_path ${output_path}
 
 if [ $? -ne 0 ]; then
-	echo "build failed : "$?
-	exit 1
+    echo "start failed : "$?
+    exit 1
 else
-	echo "finish build model!"
+    echo "Successful start service!"
 fi
+
